backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pdfplumber==0.10.2
python-docx==1.0.1
spacy==3.7.2
transformers==4.35.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
